+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=D:\Flutter2Live\flutter"
-export "FLUTTER_APPLICATION_PATH=D:\Flutter2Live\_projects\styled_text\example"
+export "FLUTTER_ROOT=D:\FlutterDev\flutter"
+export "FLUTTER_APPLICATION_PATH=D:\FlutterDev\_projects\packages\styled_text\example"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=lib\main.dart"
 export "FLUTTER_BUILD_DIR=build"