+++ styled_text/example/ios/Flutter/flutter_export_environment.sh
@@ -1,12 +1,12 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=D:\Flutter2Live\flutter"
-export "FLUTTER_APPLICATION_PATH=D:\Flutter2Live\_projects\styled_text\_temp\3.3.4"
+export "FLUTTER_ROOT=D:\FlutterDev\flutter"
+export "FLUTTER_APPLICATION_PATH=D:\FlutterDev\_projects\packages\styled_text\styled_text\example"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_TARGET=lib\main.dart"
 export "FLUTTER_BUILD_DIR=build"
-export "FLUTTER_BUILD_NAME=1.0.0"
-export "FLUTTER_BUILD_NUMBER=1"
+export "FLUTTER_BUILD_NAME=2.0.0"
+export "FLUTTER_BUILD_NUMBER=2.0.0"
 export "DART_OBFUSCATION=false"
 export "TRACK_WIDGET_CREATION=true"
 export "TREE_SHAKE_ICONS=false"